Set lastModified already on create to be consistent with v2
As Robert mentioned yesterday, v2 and v3 might use the same database for the testing in the beginning, therefore we shouldn't want to have NULL values in this database, as long as we don't tell v2 how to handle those. That might not be perfect, but as Will said this JPA implementation will be the v2-compatibility version - not the actual v3 example implementation. The real v3 one could then use real timestamps instead of long, populate lastModified only on modifications etc.